+++ CaptureHelper.swift
@@ -140,7 +140,7 @@ public protocol CaptureHelperDeviceButtonsDelegate : CaptureHelperDelegate {
 /// coming from Capture
 public protocol CaptureHelperAllDelegate : CaptureHelperErrorDelegate, CaptureHelperDevicePresenceDelegate,
     CaptureHelperDevicePowerDelegate, CaptureHelperDeviceDecodedDataDelegate,
-CaptureHelperDeviceButtonsDelegate {
+CaptureHelperDeviceButtonsDelegate, CaptureHelperDeviceManagerPresenceDelegate {
 
 }
 
@@ -871,6 +871,15 @@ public class CaptureHelper : NSObject, SKTCaptureDelegate {
                 } else {
                     self.capture = nil;
                 }
+                
+            
+                if let bundle = appInfo.mainBundle {
+                    if self.capture?.canConnectToBarcodeScanners(with: bundle) == false {
+                       print("\nIf your application uses a Socket Mobile barcode scanner, you need to add a \"UISupportedExternalAccessoryProtocols\" Array to your Info.plist with \"com.socketmobile.chs\" as an Item in that Array. Otherwise the scanner won't connect to the application.\n")
+                   }
+                }
+               
+                
                 self.callCompletion(withResult:result, withCompletion:completion)
             })
         } else {